Updated to adapter 0.2 and tagging 0.3 and removed Doctrine test
