+++ composer.json
@@ -27,16 +27,16 @@
     ],
     "require":
     {
-        "psr/cache":            "1.0.0",
-        "cache/adapter-common": "^0.1",
-        "cache/taggable-cache": "^0.2",
+        "psr/cache":            "1.0",
+        "cache/adapter-common": "^0.2",
+        "cache/taggable-cache": "^0.3",
         "doctrine/cache":       "^1.3"
     },
     "require-dev":
     {
         "phpunit/phpunit":         "^5.1|^4.0",
         "mockery/mockery":         "^0.9.4",
-        "cache/integration-tests": "dev-master"
+        "cache/integration-tests": "^0.7"
     },
     "provide":
     {
